+++ ReleaseNotes.md
@@ -15,9 +15,11 @@ Here come the actual release notes:
 ## Features
  - #201 KPI Accounted Documents
  - #212 MRPProductInfo display Conference flag with sys config
-
+ - #226 show accounting currency in open items report
+ 
 ## Fixes
  - #153 PaymentRule = S in Invoice
+ - #220 Do not load pricelist and pricelist version  on login
 
 # metasfresh 4.27.26 (2016-29, RC)